backend/requirements.txt
asgiref==3.5.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Django==4.0.4
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
Flask==2.0.2
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==9.1.0
pytz==2022.1
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.7
Werkzeug==2.0.2
